Fix failure with gcc.c-torture/execute/ashrdi-1.c -Os execution test
The following tests fail with -Os option because a shift instruction
in a branch delay slot gets replaced with multiple instructions when
the processor has no barrel shifter. This fix addresses the problem
by marking the responsible instruction pattern as type multi preventing
it from being placed in a delay slot.

>  gcc.c-torture/execute/ashrdi-1.c   -Os  execution test
>  gcc.c-torture/execute/pr40057.c   -Os  execution test
>  gcc.c-torture/execute/pr79121.c   -Os  execution test
>  gcc.c-torture/execute/pr82524.c   -Os  execution test
>  c-c++-common/torture/vector-compare-1.c   -Os  execution test
>  gcc.dg/torture/vec-cvt-1.c   -Os  execution test

These tests pass with this fix.

Signed-off-by: Gopi Kumar Bulusu <gopi@sankhya.com>